Add files from last year's failed attempt

+++ Filter.h
@@ -12,16 +12,16 @@ public:
     Filter(unsigned size);
     ~Filter();
 
-    int filter(int newVal);
-    int getFilteredVal();
-    void reset(int fillVal = 0);
+    double filter(double newVal);
+    double getFilteredVal();
+    void reset(double fillVal = 0);
 
 private:
     unsigned bufferSize;  //size of the input buffer
-    int* buffer;          //buffer of input values
+    double* buffer;          //buffer of input values
     unsigned curIndex = 0;     //index in the input values buffer
-    int filteredVal = 0;  //filtered result
-    long total = 0;       //running total of the values
+    double filteredVal = 0;  //filtered result
+    double total = 0;       //running total of the values
 };
 
 #endif
